+++ service/elasticache/examples_test.go
@@ -441,6 +441,10 @@ func ExampleElastiCache_CreateReplicationGroup_shared00() {
 				fmt.Println(elasticache.ErrCodeTagQuotaPerResourceExceeded, aerr.Error())
 			case elasticache.ErrCodeNodeGroupsPerReplicationGroupQuotaExceededFault:
 				fmt.Println(elasticache.ErrCodeNodeGroupsPerReplicationGroupQuotaExceededFault, aerr.Error())
+			case elasticache.ErrCodeGlobalReplicationGroupNotFoundFault:
+				fmt.Println(elasticache.ErrCodeGlobalReplicationGroupNotFoundFault, aerr.Error())
+			case elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault:
+				fmt.Println(elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault, aerr.Error())
 			case elasticache.ErrCodeInvalidParameterValueException:
 				fmt.Println(elasticache.ErrCodeInvalidParameterValueException, aerr.Error())
 			case elasticache.ErrCodeInvalidParameterCombinationException:
@@ -526,6 +530,10 @@ func ExampleElastiCache_CreateReplicationGroup_shared01() {
 				fmt.Println(elasticache.ErrCodeTagQuotaPerResourceExceeded, aerr.Error())
 			case elasticache.ErrCodeNodeGroupsPerReplicationGroupQuotaExceededFault:
 				fmt.Println(elasticache.ErrCodeNodeGroupsPerReplicationGroupQuotaExceededFault, aerr.Error())
+			case elasticache.ErrCodeGlobalReplicationGroupNotFoundFault:
+				fmt.Println(elasticache.ErrCodeGlobalReplicationGroupNotFoundFault, aerr.Error())
+			case elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault:
+				fmt.Println(elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault, aerr.Error())
 			case elasticache.ErrCodeInvalidParameterValueException:
 				fmt.Println(elasticache.ErrCodeInvalidParameterValueException, aerr.Error())
 			case elasticache.ErrCodeInvalidParameterCombinationException:
@@ -1653,6 +1661,8 @@ func ExampleElastiCache_ModifyCacheParameterGroup_shared00() {
 				fmt.Println(elasticache.ErrCodeInvalidParameterValueException, aerr.Error())
 			case elasticache.ErrCodeInvalidParameterCombinationException:
 				fmt.Println(elasticache.ErrCodeInvalidParameterCombinationException, aerr.Error())
+			case elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault:
+				fmt.Println(elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault, aerr.Error())
 			default:
 				fmt.Println(aerr.Error())
 			}
@@ -1900,6 +1910,8 @@ func ExampleElastiCache_ResetCacheParameterGroup_shared00() {
 				fmt.Println(elasticache.ErrCodeInvalidParameterValueException, aerr.Error())
 			case elasticache.ErrCodeInvalidParameterCombinationException:
 				fmt.Println(elasticache.ErrCodeInvalidParameterCombinationException, aerr.Error())
+			case elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault:
+				fmt.Println(elasticache.ErrCodeInvalidGlobalReplicationGroupStateFault, aerr.Error())
 			default:
 				fmt.Println(aerr.Error())
 			}